split config tests into files
